enable Checkout.modifyThenRevert
Summary:
At the end of the test, the test assumes that an old uncommited version of
the file can be retrieved thanks to the files living in the overlay. On
Windows, the files are written directly onto the Projected FS cache, and thus
the old version of the file cannot be retrieved.

Reviewed By: pkaush

Differential Revision: D21193209

fbshipit-source-id: a4db6fb2bfc68817601725f97a6d47f2319b83c7