remotenames: normalize infinitepush/ to default/
Summary:
Similar to part of D20804854, normalize infinitepush/ to default/ in
remotenames code paths.

This might not be strictly necessary, but we never want to write out
infinitepush/ remote bookmarks and we have seen that OnDemand writes out
infinitepush/ remote bookmarks (which got synced to commit cloud and caused
user complaints) recently.

Reviewed By: DurhamG

Differential Revision: D21233658

fbshipit-source-id: cf8c12ac40348c7b475a7f3d5d5223775491650e