mononoke_api: fix for acl check in tests
Summary:
This is a fix after D21067809 broke the mononoke_api unittests and was not caught by sandcastle.

Prior to this change if `ctx.identities` were None then the permission was unconditionally denied, even when the permission checker is set to always_allow. After this change if the identities is None then the permission_checker is called with empty identity set thus making unit tests work properly.

Reviewed By: krallin

Differential Revision: D21324038

fbshipit-source-id: 1edd5fd2af6731d43cab06324b4d9fc02882c09e